# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FamilySpace API - private family communication and experiences platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /experiences:
    get:
      operationId: listExperiences
      tags: [experiences]
      summary: List experience recommendations
      parameters:
        - name: timeframe
          in: query
          schema:
            type: string
            enum: [right-now, this-week, this-weekend, this-month, this-season, this-year, surprise]
        - name: participants
          in: query
          schema:
            type: string
        - name: maxCost
          in: query
          schema:
            type: number
        - name: maxDistance
          in: query
          schema:
            type: number
        - name: energyLevel
          in: query
          schema:
            type: string
            enum: [low, medium, high]
        - name: setting
          in: query
          schema:
            type: string
            enum: [home, indoor, outdoor, any]
        - name: needsChildcare
          in: query
          schema:
            type: boolean
        - name: surpriseMode
          in: query
          schema:
            type: boolean
        - name: savedOnly
          in: query
          schema:
            type: boolean
        - name: shortlistedOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of experiences
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Experience"
  /experiences/search:
    get:
      operationId: searchExperiences
      tags: [experiences]
      summary: Natural language experience search
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: viewerId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Search results with parsed filters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExperienceSearchResult"
  /experiences/{id}:
    get:
      operationId: getExperience
      tags: [experiences]
      summary: Get a single experience
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Experience detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"
        "404":
          description: Not found
  /experiences/{id}/save:
    post:
      operationId: saveExperience
      tags: [experiences]
      summary: Save or unsave an experience
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperienceActionInput"
      responses:
        "200":
          description: Updated experience
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"
  /experiences/{id}/hide:
    post:
      operationId: hideExperience
      tags: [experiences]
      summary: Hide or unhide an experience
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperienceActionInput"
      responses:
        "200":
          description: Updated experience
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"
  /experiences/{id}/shortlist:
    post:
      operationId: shortlistExperience
      tags: [experiences]
      summary: Add or remove experience from shortlist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExperienceActionInput"
      responses:
        "200":
          description: Updated experience
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experience"
...
  /calendar-events:
    get:
      operationId: listCalendarEvents
      tags: [calendar-events]
      summary: List calendar events
      parameters:
        - name: from
          in: query
          schema:
            type: string
        - name: to
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of calendar events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
    post:
      operationId: createCalendarEvent
      tags: [calendar-events]
      summary: Add event to shared calendar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalendarEventInput"
      responses:
        "201":
          description: Created calendar event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
...
components:
  schemas:
...
    Experience:
      type: object
      required: [id, title, description, imageUrl, participants, timeframes, durationMinutes, costEstimate, energyLevel, setting, planningEffort, isSaved, isHidden, isShortlisted, whyItFits, privacySafeSource, demoDataNote]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        participants:
          type: array
          items:
            type: string
        participantIds:
          type: array
          items:
            type: string
        timeframes:
          type: array
          items:
            type: string
            enum: [right-now, this-week, this-weekend, this-month, this-season, this-year, surprise]
        durationMinutes:
          type: integer
        costEstimate:
          type: number
        costBreakdown:
          type: ["string", "null"]
        energyLevel:
          type: string
          enum: [low, medium, high]
        setting:
          type: string
          enum: [home, indoor, outdoor, any]
        planningEffort:
          type: string
          enum: [none, minimal, moderate, significant]
        needsChildcare:
          type: boolean
        needsReservation:
          type: boolean
        weatherConsideration:
          type: ["string", "null"]
        accessibilityNotes:
          type: ["string", "null"]
        isSaved:
          type: boolean
        isHidden:
          type: boolean
        isShortlisted:
          type: boolean
        whyItFits:
          type: string
        privacySafeSource:
          type: string
        demoDataNote:
          type: string
        distanceMiles:
          type: ["number", "null"]
        travelTimeMinutes:
          type: ["integer", "null"]
        crowdLevel:
          type: ["string", "null"]
        surpriseEligible:
          type: boolean
        category:
          type: string
        tags:
          type: array
          items:
            type: string
        demoAvailability:
          type: ["string", "null"]
    ExperienceActionInput:
      type: object
      required: [value]
      properties:
        value:
          type: boolean
        memberId:
          type: string
    ExperienceSearchResult:
      type: object
      required: [experiences, parsedFilters, query]
      properties:
        query:
          type: string
        parsedFilters:
          $ref: "#/components/schemas/ParsedSearchFilters"
        experiences:
          type: array
          items:
            $ref: "#/components/schemas/Experience"
    ParsedSearchFilters:
      type: object
      properties:
        participants:
          type: array
          items:
            type: string
        dayOfWeek:
          type: ["string", "null"]
        timeOfDay:
          type: ["string", "null"]
        maxCost:
          type: ["number", "null"]
        maxDistance:
          type: ["number", "null"]
        crowdLevel:
          type: ["string", "null"]
        energyLevel:
          type: ["string", "null"]
        setting:
          type: ["string", "null"]
        month:
          type: ["string", "null"]
        surpriseMode:
          type: ["boolean", "null"]
        durationHours:
          type: ["number", "null"]
        chips:
          type: array
          items:
            $ref: "#/components/schemas/FilterChip"
    FilterChip:
      type: object
      required: [key, label, value]
      properties:
        key:
          type: string
        label:
          type: string
        value:
          type: string
        editable:
          type: boolean
...
    CalendarEvent:
      type: object
      required: [id, title, date, participantIds, createdAt]
      properties:
        id:
          type: string
        title:
          type: string
        date:
          type: string
        time:
          type: ["string", "null"]
        durationMinutes:
          type: ["integer", "null"]
        participantIds:
          type: array
          items:
            type: string
        experienceId:
          type: ["string", "null"]
        invitationId:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        createdAt:
          type: string
    CalendarEventInput:
      type: object
      required: [title, date, participantIds]
      properties:
        title:
          type: string
        date:
          type: string
        time:
          type: string
        durationMinutes:
          type: integer
        participantIds:
          type: array
          items:
            type: string
        experienceId:
          type: string
        invitationId:
          type: string
        notes:
          type: string
        location:
          type: string
...